'modules/program.sh.m4':         fixed        an         error        in "mbfl_program_validate_declared".
"modules/main.sh.m4':  added  private main  function  to let  predefined
options override  the script main  function; this lets the  script parse
all  the  command  line  options   and  then  invoke  the  private  main
function. Useful,  for example,  when "--validate-programs" is  used: if
the  "--verbose" option  is placed  after  it, it  is recognised  before
entering  the  validate function;  before  it  was  ignored because  the
validate function  was invoked immediately, without  letting the getopts
module to parse it. 


* Jan 29 2005, 08:50:13: committing revision


git-archimport-id: marco@rapitore.luna--devel-2004/mbfl--stable--1.0--patch-6